requirements.txt
aiohttp>=2.2.3
async-timeout>=1.2.1
chardet>=3.0.4
discord.py>=1.0.0a1275
multidict>=3.1.1
websockets>=3.3
yarl>=0.11.0
